pyproject.toml
[tool.poetry]
name = "djoek"
version = "0.1.0"
description = ""
authors = ["Ingmar Steen <iksteen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
peewee-async = "^0.9.1"
fastapi = "^0.103.2"
aiofiles = "^23.2.1"
httpx = "^0.25.0"
aiopg = "^1.0.0"
uvicorn = {version="^0.23.2", extras = ["standard"]}
python-dotenv = "^1.0.0"
python-jose = {version = "^3.1.0", extras = ["cryptography"]}
multidict = "^6.0.2"
mutagen = "^1.44.0"
isodate = "^0.6.0"
cachetools = "^5.2.0"
yt-dlp = "^2023.9.24"

[tool.poetry.dev-dependencies]
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
